switched Inventory class public function returnItemSlot to take 2 int inputs instead of 2 uint8_t

+++ dchu.h
@@ -65,7 +65,7 @@ class Inventory
 
         //Getters:
         void getInventory();
-        item returnItemSlot(uint8_t, uint8_t);
+        item returnItemSlot(int, int);
 };
 
 //ECS: Inventory Management System
